pyproject.toml
[project]
name = "parking-services"
version = "0.1.0"
description = "ML project with training and web server"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
training = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tensorboard>=2.14.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
]

django = [
    "django>=5.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-api-key>=3.1.0",
    "django-cors-headers>=4.3.0",
    "gunicorn>=21.2.0",
    "django-environ>=0.12.0",
    "psycopg2-binary>=2.9.9",  # для PostgreSQL
    "pillow>=10.0.0",  # для ImageField
    "celery>=5.3.0",  # для фоновых задач (опционально)
    "redis>=5.0.0",  # для кеширования (опционально)
]

orangepi = [
    "ultralytics>=8.0.0",
    "opencv-python-headless>=4.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["parking_services"]

